Refactor types, cleanup
