+++ src/typescript/ProjectInfo.ts
@@ -1,0 +1,14 @@
+import { TargetFrameworkInfo } from "./TargetFrameworkInfo";
+
+
+export class ProjectInfo {
+
+	Name: string;
+	Path: string;
+	AssemblyName: string;
+	TargetFrameworks: TargetFrameworkInfo[];
+	IsExe: boolean;
+	Configurations: string[];
+	Platforms: string[];
+	OutputPath: string;
+}
