+++ src/typescript/SolutionInfo.ts
@@ -1,0 +1,6 @@
+import { ProjectInfo } from "./ProjectInfo";
+
+
+export class SolutionInfo {
+	Projects: ProjectInfo[];
+}
