+++ src/typescript/TargetFrameworkInfo.ts
@@ -1,0 +1,8 @@
+
+export class TargetFrameworkInfo {
+
+	FullName: string;
+	Platform: string;
+	Version: string;
+	PlatformVersion: string;
+}
