+++ src/typescript/WorkspaceInfo.ts
@@ -1,0 +1,6 @@
+import { SolutionInfo } from "./SolutionInfo";
+
+
+export class WorkspaceInfo {
+	Solution: SolutionInfo;
+}
